Moved winpcapy to modules
