+++ .coveragerc
@@ -8,3 +8,4 @@ omit =
     scapy/tools/*
     # Libraries
     scapy/modules/six.py
+	scapy/modules/winpcapy.py
